1.完成离职审核员台账业务逻辑及导出功能； 2.添加事业部变动审核员台账分页搜索功能。

+++ platform-pojo/src/main/java/com/platform/dto/factories/FactoryCreateDTO.java
@@ -19,5 +19,5 @@ public class FactoryCreateDTO implements Serializable {
     @Schema(description = "事业部编号")
     private Long buId;
     @Schema(description = "星级")
-    private int level;
+    private Long level;
 }
